fix: quick auth fixes
