+++ .env.example
@@ -2,3 +2,7 @@
 # https://app.supabase.com/project/_/settings/api
 NEXT_PUBLIC_SUPABASE_URL=your-project-url
 NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY=your-publishable-or-anon-key
+
+NEXT_PUBLIC_SITE_URL=https://yourdomain.com
+
+DATABASE_URL=postgresql://user:password@host:port/postgres
